+++ lib/absinthe/plug/batch/runner.ex
@@ -1,5 +1,4 @@
 defmodule Absinthe.Plug.Batch.Runner do
-
   @moduledoc false
 
   alias Absinthe.Plug.Request
@@ -9,10 +8,11 @@ defmodule Absinthe.Plug.Batch.Runner do
 
     queries = prepare(queries)
 
-    {valid_queries, invalid_queries} = Enum.split_with(queries, fn
-      {:ok, _, _, _} -> true
-      {:error, _, _, _} -> false
-    end)
+    {valid_queries, invalid_queries} =
+      Enum.split_with(queries, fn
+        {:ok, _, _, _} -> true
+        {:error, _, _, _} -> false
+      end)
 
     valid_results = build_valid_results(valid_queries, config.schema_mod)
     invalid_results = build_invalid_results(invalid_queries)
@@ -24,19 +24,21 @@ defmodule Absinthe.Plug.Batch.Runner do
   end
 
   defp restore_order(valid_results, invalid_results) do
-    valid_results ++ invalid_results
+    (valid_results ++ invalid_results)
     |> Enum.sort_by(fn {i, _q} -> i end)
     |> Enum.map(fn {_i, q} -> q end)
   end
 
   defp build_valid_results(valid_queries, schema) do
-    blueprints = Enum.map(valid_queries, fn
-      {:ok, bp, _query, _index} -> bp
-    end)
+    blueprints =
+      Enum.map(valid_queries, fn
+        {:ok, bp, _query, _index} -> bp
+      end)
 
-    querys_and_indices = Enum.map(valid_queries, fn
-      {:ok, _bp, query, index} -> {query, index}
-    end)
+    querys_and_indices =
+      Enum.map(valid_queries, fn
+        {:ok, _bp, query, index} -> {query, index}
+      end)
 
     blueprints
     |> Absinthe.Pipeline.BatchResolver.run(schema: schema)
@@ -56,6 +58,7 @@ defmodule Absinthe.Plug.Batch.Runner do
     case Absinthe.Pipeline.run(bp, result_pipeline(query)) do
       {:ok, bp, _} ->
         bp
+
       _ ->
         %{result: %{errors: ["could not produce a valid JSON result"]}}
     end
@@ -68,9 +71,11 @@ defmodule Absinthe.Plug.Batch.Runner do
           case bp.execution.validation_errors do
             [] ->
               {:ok, bp, query, i}
+
             _ ->
               {:error, bp, query, i}
           end
+
         {:error, bp, _} ->
           {:error, bp, query, i}
       end