+++ lib/absinthe/plug/graphiql.ex
@@ -129,32 +129,44 @@ defmodule Absinthe.Plug.GraphiQL do
 
   @graphiql_template_path Path.join(__DIR__, "graphiql")
 
-  EEx.function_from_file :defp, :graphiql_html, Path.join(@graphiql_template_path, "graphiql.html.eex"),
+  EEx.function_from_file(
+    :defp,
+    :graphiql_html,
+    Path.join(@graphiql_template_path, "graphiql.html.eex"),
     [:query_string, :variables_string, :result_string, :socket_url, :assets]
+  )
 
-  EEx.function_from_file :defp, :graphiql_workspace_html, Path.join(@graphiql_template_path, "graphiql_workspace.html.eex"),
+  EEx.function_from_file(
+    :defp,
+    :graphiql_workspace_html,
+    Path.join(@graphiql_template_path, "graphiql_workspace.html.eex"),
     [:query_string, :variables_string, :default_headers, :default_url, :socket_url, :assets]
+  )
 
-  EEx.function_from_file :defp, :graphiql_playground_html, Path.join(@graphiql_template_path, "graphiql_playground.html.eex"),
-  [:default_url, :socket_url, :assets]
+  EEx.function_from_file(
+    :defp,
+    :graphiql_playground_html,
+    Path.join(@graphiql_template_path, "graphiql_playground.html.eex"),
+    [:default_url, :socket_url, :assets]
+  )
 
   @behaviour Plug
 
   import Plug.Conn
 
   @type opts :: [
-    schema: atom,
-    adapter: atom,
-    path: binary,
-    context: map,
-    json_codec: atom | {atom, Keyword.t},
-    interface: :playground | :advanced | :simple,
-    default_headers: {module, atom},
-    default_url: binary,
-    assets: Keyword.t,
-    socket: module,
-    socket_url: binary,
-  ]
+          schema: atom,
+          adapter: atom,
+          path: binary,
+          context: map,
+          json_codec: atom | {atom, Keyword.t()},
+          interface: :playground | :advanced | :simple,
+          default_headers: {module, atom},
+          default_url: binary,
+          assets: Keyword.t(),
+          socket: module,
+          socket_url: binary
+        ]
 
   @doc false
   @spec init(opts :: opts) :: map
@@ -162,13 +174,14 @@ defmodule Absinthe.Plug.GraphiQL do
     assets = Absinthe.Plug.GraphiQL.Assets.get_assets()
 
     opts
-    |> Absinthe.Plug.init
+    |> Absinthe.Plug.init()
     |> Map.put(:interface, Keyword.get(opts, :interface) || :advanced)
     |> Map.put(:default_headers, Keyword.get(opts, :default_headers))
     |> Map.put(:default_url, Keyword.get(opts, :default_url))
     |> Map.put(:assets, assets)
     |> Map.put(:socket, Keyword.get(opts, :socket))
     |> Map.put(:socket_url, Keyword.get(opts, :socket_url))
+    |> Map.put(:default_query, Keyword.get(opts, :default_query, ""))
     |> set_pipeline
   end
 
@@ -182,10 +195,11 @@ defmodule Absinthe.Plug.GraphiQL do
 
   defp html?(conn) do
     Plug.Conn.get_req_header(conn, "accept")
-    |> List.first
+    |> List.first()
     |> case do
       string when is_binary(string) ->
         String.contains?(string, "text/html")
+
       _ ->
         false
     end
@@ -202,40 +216,47 @@ defmodule Absinthe.Plug.GraphiQL do
          {:process, request} <- select_mode(request),
          {:ok, request} <- Absinthe.Plug.ensure_processable(request, config),
          :ok <- Absinthe.Plug.Request.log(request, config.log_level) do
-
       conn_info = %{
-        conn_private: (conn.private[:absinthe] || %{}) |> Map.put(:http_method, conn.method),
+        conn_private: (conn.private[:absinthe] || %{}) |> Map.put(:http_method, conn.method)
       }
 
       {conn, result} = Absinthe.Plug.run_request(request, conn, conn_info, config)
+
       case result do
         {:ok, result} ->
-          query = hd(request.queries) # GraphiQL doesn't batch requests, so the first query is the only one
+          # GraphiQL doesn't batch requests, so the first query is the only one
+          query = hd(request.queries)
           {:ok, conn, result, query.variables, query.document || ""}
+
         {:error, {:http_method, _}, _} ->
           query = hd(request.queries)
           {:http_method_error, query.variables, query.document || ""}
-        other -> other
+
+        other ->
+          other
       end
     end
     |> case do
       {:ok, conn, result, variables, query} ->
         query = query |> js_escape
 
-        var_string = variables
-        |> config.json_codec.module.encode!(pretty: true)
-        |> js_escape
+        var_string =
+          variables
+          |> config.json_codec.module.encode!(pretty: true)
+          |> js_escape
 
+        result =
+          result
+          |> config.json_codec.module.encode!(pretty: true)
+          |> js_escape
 
-        result = result
-        |> config.json_codec.module.encode!(pretty: true)
-        |> js_escape
-
-        config = %{
-          query: query,
-          var_string: var_string,
-          result: result,
-        } |> Map.merge(config)
+        config =
+          %{
+            query: query,
+            var_string: var_string,
+            result: result
+          }
+          |> Map.merge(config)
 
         conn
         |> render_interface(interface, config)
@@ -245,7 +266,6 @@ defmodule Absinthe.Plug.GraphiQL do
         |> send_resp(400, msg)
 
       :start_interface ->
-
         conn
         |> render_interface(interface, config)
 
@@ -257,10 +277,12 @@ defmodule Absinthe.Plug.GraphiQL do
           |> config.json_codec.module.encode!(pretty: true)
           |> js_escape
 
-        config = %{
-          query: query,
-          var_string: var_string,
-        } |> Map.merge(config)
+        config =
+          %{
+            query: query,
+            var_string: var_string
+          }
+          |> Map.merge(config)
 
         conn
         |> render_interface(interface, config)
@@ -268,7 +290,6 @@ defmodule Absinthe.Plug.GraphiQL do
       {:error, error, _} when is_binary(error) ->
         conn
         |> send_resp(500, error)
-
     end
   end
 
@@ -283,15 +304,19 @@ defmodule Absinthe.Plug.GraphiQL do
     {module, fun} = config.additional_pipeline
 
     apply(module, fun, [config, opts])
-    |> Absinthe.Pipeline.insert_after(Absinthe.Phase.Document.CurrentOperation,
+    |> Absinthe.Pipeline.insert_after(
+      Absinthe.Phase.Document.CurrentOperation,
       [
-        Absinthe.GraphiQL.Validation.NoSubscriptionOnHTTP,
+        Absinthe.GraphiQL.Validation.NoSubscriptionOnHTTP
       ]
     )
   end
 
-  @spec select_mode(request :: Absinthe.Plug.Request.t) :: :start_interface | {:process, Absinthe.Plug.Request.t}
-  defp select_mode(%{queries: [%Absinthe.Plug.Request.Query{document: nil}]}), do: :start_interface
+  @spec select_mode(request :: Absinthe.Plug.Request.t()) ::
+          :start_interface | {:process, Absinthe.Plug.Request.t()}
+  defp select_mode(%{queries: [%Absinthe.Plug.Request.Query{document: nil}]}),
+    do: :start_interface
+
   defp select_mode(request), do: {:process, request}
 
   defp find_socket_path(conn, socket) do
@@ -308,40 +333,60 @@ defmodule Absinthe.Plug.GraphiQL do
     end
   end
 
-  @render_defaults %{query: "", var_string: "", results: ""}
+  @render_defaults %{var_string: "", results: ""}
 
-  @spec render_interface(conn :: Conn.t, interface :: :advanced | :simple, opts :: Keyword.t) :: Conn.t
+  @spec render_interface(Plug.Conn.t(), :advanced | :simple | :playground, map()) ::
+          Plug.Conn.t()
   defp render_interface(conn, interface, opts)
+
   defp render_interface(conn, :simple, opts) do
-    opts = Map.merge(@render_defaults, opts)
+    opts = opts_with_default(opts)
 
     graphiql_html(
-      opts[:query], opts[:var_string], opts[:result], opts[:socket_url], opts[:assets]
+      opts[:query],
+      opts[:var_string],
+      opts[:result],
+      opts[:socket_url],
+      opts[:assets]
     )
     |> rendered(conn)
   end
+
   defp render_interface(conn, :advanced, opts) do
-    opts = Map.merge(@render_defaults, opts)
+    opts = opts_with_default(opts)
 
     graphiql_workspace_html(
-      opts[:query], opts[:var_string], opts[:default_headers],
-      default_url(opts[:default_url]), opts[:socket_url], opts[:assets]
+      opts[:query],
+      opts[:var_string],
+      opts[:default_headers],
+      default_url(opts[:default_url]),
+      opts[:socket_url],
+      opts[:assets]
     )
     |> rendered(conn)
   end
+
   defp render_interface(conn, :playground, opts) do
-    opts = Map.merge(@render_defaults, opts)
+    opts = opts_with_default(opts)
 
     graphiql_playground_html(
-      default_url(opts[:default_url]), opts[:socket_url], opts[:assets]
+      default_url(opts[:default_url]),
+      opts[:socket_url],
+      opts[:assets]
     )
     |> rendered(conn)
+  end
+
+  defp opts_with_default(opts) do
+    defaults = Map.put(@render_defaults, :query, opts[:default_query])
+
+    Map.merge(defaults, opts)
   end
 
   defp default_url(nil), do: "window.location.origin + window.location.pathname"
   defp default_url(url), do: "'#{url}'"
 
-  @spec rendered(String.t, Plug.Conn.t) :: Conn.t
+  @spec rendered(String.t(), Plug.Conn.t()) :: Plug.Conn.t()
   defp rendered(html, conn) do
     conn
     |> put_resp_content_type("text/html")
@@ -368,7 +413,7 @@ defmodule Absinthe.Plug.GraphiQL do
         Map.put(config, :default_headers, header_string)
 
       val ->
-        raise "invalid default headers: #{inspect val}"
+        raise "invalid default headers: #{inspect(val)}"
     end
   end
 
@@ -380,10 +425,12 @@ defmodule Absinthe.Plug.GraphiQL do
     case get_config_val(config, key, conn) do
       nil ->
         config
+
       val when is_binary(val) ->
         Map.put(config, key, val)
+
       val ->
-        raise "invalid #{key}: #{inspect val}"
+        raise "invalid #{key}: #{inspect(val)}"
     end
   end
 
@@ -391,11 +438,16 @@ defmodule Absinthe.Plug.GraphiQL do
     case Map.get(config, key) do
       {module, fun} when is_atom(fun) ->
         case function_arity(module, fun) do
-          1 -> apply(module, fun, [conn])
-          0 -> apply(module, fun, [])
+          1 ->
+            apply(module, fun, [conn])
+
+          0 ->
+            apply(module, fun, [])
+
           _ ->
             raise "function for #{key}: {#{module}, #{fun}} is not exported with arity 1 or 0"
         end
+
       val ->
         val
     end
@@ -414,10 +466,11 @@ defmodule Absinthe.Plug.GraphiQL do
       else
         _ -> "''"
       end
+
     %{config | socket_url: url}
   end
+
   defp normalize_socket_url(%{socket_url: url} = config, _) do
     %{config | socket_url: "'#{url}'"}
   end
-
 end