+++ lib/absinthe/plug/graphiql/assets.ex
@@ -8,42 +8,50 @@ defmodule Absinthe.Plug.GraphiQL.Assets do
     local_url_path: "/absinthe_graphiql",
     local_directory: "priv/static/absinthe_graphiql",
     local_source: ":package/:alias",
-    remote_source: "https://cdn.jsdelivr.net/npm/:package@:version/:file",
+    remote_source: "https://cdn.jsdelivr.net/npm/:package@:version/:file"
   ]
 
   @react_version "15.6.1"
 
   @assets [
-    {"whatwg-fetch", "2.0.3", [
-      {"fetch.min.js", "fetch.js"},
-    ]},
-    {"react", @react_version, [
-      {"dist/react.min.js", "react.js"},
-    ]},
-    {"react-dom", @react_version, [
-      {"dist/react-dom.min.js", "react-dom.js"},
-    ]},
-    {"bootstrap", "3.3.7", [
-      {"dist/fonts/glyphicons-halflings-regular.eot", "fonts/glyphicons-halflings-regular.eot"},
-      {"dist/fonts/glyphicons-halflings-regular.ttf", "fonts/glyphicons-halflings-regular.ttf"},
-      {"dist/fonts/glyphicons-halflings-regular.woff2", "fonts/glyphicons-halflings-regular.woff2"},
-      {"dist/fonts/glyphicons-halflings-regular.svg", "fonts/glyphicons-halflings-regular.svg"},
-      {"dist/css/bootstrap.min.css", "css/bootstrap.css"},
-    ]},
-    {"graphiql", "0.11.10", [
-      "graphiql.css",
-      {"graphiql.min.js", "graphiql.js"},
-    ]},
-    {"graphiql-workspace", "1.1.4", [
-      "graphiql-workspace.css",
-      {"graphiql-workspace.min.js", "graphiql-workspace.js"}
-    ]},
+    {"whatwg-fetch", "2.0.3",
+     [
+       {"fetch.min.js", "fetch.js"}
+     ]},
+    {"react", @react_version,
+     [
+       {"dist/react.min.js", "react.js"}
+     ]},
+    {"react-dom", @react_version,
+     [
+       {"dist/react-dom.min.js", "react-dom.js"}
+     ]},
+    {"bootstrap", "3.3.7",
+     [
+       {"dist/fonts/glyphicons-halflings-regular.eot", "fonts/glyphicons-halflings-regular.eot"},
+       {"dist/fonts/glyphicons-halflings-regular.ttf", "fonts/glyphicons-halflings-regular.ttf"},
+       {"dist/fonts/glyphicons-halflings-regular.woff2",
+        "fonts/glyphicons-halflings-regular.woff2"},
+       {"dist/fonts/glyphicons-halflings-regular.svg", "fonts/glyphicons-halflings-regular.svg"},
+       {"dist/css/bootstrap.min.css", "css/bootstrap.css"}
+     ]},
+    {"graphiql", "0.11.10",
+     [
+       "graphiql.css",
+       {"graphiql.min.js", "graphiql.js"}
+     ]},
+    {"graphiql-workspace", "1.1.4",
+     [
+       "graphiql-workspace.css",
+       {"graphiql-workspace.min.js", "graphiql-workspace.js"}
+     ]},
     # Used by graphql-playground
-    {"typeface-source-code-pro", "0.0.44", [
-      {"index.css", "index.css"},
-      {"files/source-code-pro-latin-400.woff2", "files/source-code-pro-latin-400.woff2"},
-      {"files/source-code-pro-latin-700.woff2", "files/source-code-pro-latin-700.woff2"},
-    ]},
+    {"typeface-source-code-pro", "0.0.44",
+     [
+       {"index.css", "index.css"},
+       {"files/source-code-pro-latin-400.woff2", "files/source-code-pro-latin-400.woff2"},
+       {"files/source-code-pro-latin-700.woff2", "files/source-code-pro-latin-700.woff2"}
+     ]},
     # Used by graphql-playground
     {"typeface-open-sans", "0.0.44", [
       {"index.css", "index.css"},
@@ -64,6 +72,7 @@ defmodule Absinthe.Plug.GraphiQL.Assets do
     case @config do
       nil ->
         @default_config
+
       config ->
         Keyword.merge(@default_config, Keyword.get(config, :assets, []))
     end
@@ -83,10 +92,13 @@ defmodule Absinthe.Plug.GraphiQL.Assets do
   def get_remote_asset_mappings do
     reduce_assets(
       [],
-      &(&2 ++ [{
-        local_asset_path(&1),
-        asset_source_url(:remote, &1)
-      }])
+      &(&2 ++
+          [
+            {
+              local_asset_path(&1),
+              asset_source_url(:remote, &1)
+            }
+          ])
     )
   end
 
@@ -107,9 +119,11 @@ defmodule Absinthe.Plug.GraphiQL.Assets do
       asset_source_url(:remote, asset)
     end
   end
+
   defp asset_source_url(:local, asset) do
     Path.join(assets_config()[:local_url_path], build_asset_path(:local_source, asset))
   end
+
   defp asset_source_url(:remote, asset) do
     build_asset_path(:remote_source, asset)
   end
@@ -121,6 +135,7 @@ defmodule Absinthe.Plug.GraphiQL.Assets do
     |> String.replace(":file", real_path)
     |> String.replace(":alias", aliased_path)
   end
+
   defp build_asset_path(source, {package, version, path}) do
     build_asset_path(source, {package, version, {path, path}})
   end