+++ lib/absinthe/plug/parser.ex
@@ -11,31 +11,34 @@ defmodule Absinthe.Plug.Parser do
       plug Plug.Parsers,
         parsers: [:urlencoded, :multipart, :json, Absinthe.Plug.Parser],
         pass: ["*/*"],
-        json_decoder: Poison
+        json_decoder: Jason
 
       plug Absinthe.Plug,
         schema: MyAppWeb.Schema
   """
 
   @behaviour Plug.Parsers
-  alias Plug.Conn
 
   @doc false
   def init(opts), do: opts
 
   @doc false
   def parse(conn, "application", "graphql", _headers, opts) do
-    case Conn.read_body(conn, opts) do
+    case Plug.Conn.read_body(conn, opts) do
       {:ok, body, conn} ->
         {:ok, %{"query" => body}, conn}
+
       {:more, _data, conn} ->
         {:error, :too_large, conn}
+
       {:error, :timeout} ->
         raise Plug.TimeoutError
+
       {:error, _} ->
         raise Plug.BadRequestError
     end
   end
+
   def parse(conn, _type, _subtype, _headers, _opts) do
     {:next, conn}
   end