+++ lib/absinthe/validation/no_subscription_on_http.ex
@@ -8,10 +8,12 @@ defmodule Absinthe.GraphiQL.Validation.NoSubscriptionOnHTTP do
   def run(blueprint, _) do
     case Blueprint.current_operation(blueprint) do
       %{type: :subscription} ->
-        {:error, {:http_method, "Subscriptions cannot be run over HTTP. Please configure a websocket connection"}}
+        {:error,
+         {:http_method,
+          "Subscriptions cannot be run over HTTP. Please configure a websocket connection"}}
+
       _ ->
         {:ok, blueprint}
     end
   end
-
 end