+++ lib/mix/tasks/absinthe.plug.graphiql.assets.download.ex
@@ -6,7 +6,8 @@ defmodule Mix.Tasks.Absinthe.Plug.Graphiql.Assets.Download do
   def run(args) do
     Mix.Absinthe.Plug.GraphiQL.AssetsTask.run(args)
 
-    Application.ensure_all_started(:inets)
+    {:ok, _} = Application.ensure_all_started(:inets)
+    {:ok, _} = Application.ensure_all_started(:ssl)
 
     Absinthe.Plug.GraphiQL.Assets.get_remote_asset_mappings()
     |> Enum.map(&download_file/1)
@@ -17,13 +18,15 @@ defmodule Mix.Tasks.Absinthe.Plug.Graphiql.Assets.Download do
 
     case response do
       {{_, http_code, _}, _, body} when http_code in [200] ->
-        Mix.Generator.create_file(destination, body, [force: true])
+        Mix.Generator.create_file(destination, body, force: true)
+
       _ ->
-        Mix.raise """
+        Mix.raise("""
           Something went wrong downloading #{asset_url} Please try again.
-        """
+        """)
     end
   end
 
-  defp http_get(url), do: :httpc.request(:get, {String.to_charlist(url), []}, [], [body_format: :binary])
+  defp http_get(url),
+    do: :httpc.request(:get, {String.to_charlist(url), []}, [], body_format: :binary)
 end