+++ test/lib/absinthe/plug/document_provider/compiled_test.exs
@@ -3,33 +3,16 @@ defmodule Absinthe.Plug.DocumentProvider.CompiledTest do
   alias Absinthe.Plug.TestSchema
   alias Absinthe.Plug.DocumentProvider.Compiled
 
-  defmodule LiteralDocuments do
-    use Absinthe.Plug.DocumentProvider.Compiled
-
-    provide "1", """
-    query FooQuery($id: ID!) {
-      item(id: $id) {
-        name
-      }
-    }
-    """
-
-    provide "2", """
-    query GetUser {
-      user
-    }
-    """
-  end
-
   defmodule ExtractedDocuments do
     use Absinthe.Plug.DocumentProvider.Compiled
 
-    @fixture Path.join([File.cwd!, "test/support/fixtures/extracted_queries.json"])
+    @fixture Path.join([File.cwd!(), "test/support/fixtures/extracted_queries.json"])
 
-    provide File.read!(@fixture)
-    |> Poison.decode!
-    |> Map.new(fn {k, v} -> {v, k} end)
-
+    provide(
+      File.read!(@fixture)
+      |> Jason.decode!()
+      |> Map.new(fn {k, v} -> {v, k} end)
+    )
   end
 
   @query """
@@ -42,12 +25,17 @@ defmodule Absinthe.Plug.DocumentProvider.CompiledTest do
   @result ~s({"data":{"item":{"name":"Foo"}}})
 
   test "works using documents provided as literals" do
-    opts = Absinthe.Plug.init(schema: TestSchema, document_providers: [__MODULE__.LiteralDocuments])
+    opts =
+      Absinthe.Plug.init(
+        schema: TestSchema,
+        document_providers: [Absinthe.Plug.TestLiteralDocuments]
+      )
 
-    assert %{status: 200, resp_body: resp_body} = conn(:post, "/", %{"id" => "1", "variables" => %{"id" => "foo"}})
-    |> put_req_header("content-type", "application/graphql")
-    |> plug_parser
-    |> Absinthe.Plug.call(opts)
+    assert %{status: 200, resp_body: resp_body} =
+             conn(:post, "/", %{"id" => "1", "variables" => %{"id" => "foo"}})
+             |> put_req_header("content-type", "application/graphql")
+             |> plug_parser
+             |> Absinthe.Plug.call(opts)
 
     assert resp_body == @result
   end
@@ -60,45 +48,89 @@ defmodule Absinthe.Plug.DocumentProvider.CompiledTest do
       user
     }
     """
-    assert %{status: 200, resp_body: resp_body} = conn(:post, "/", %{"query" => query})
-    |> Absinthe.Plug.put_options(context: %{user: "Foo"})
-    |> put_req_header("content-type", "application/graphql")
-    |> plug_parser
-    |> Absinthe.Plug.call(opts)
+
+    assert %{status: 200, resp_body: resp_body} =
+             conn(:post, "/", %{"query" => query})
+             |> Absinthe.Plug.put_options(context: %{user: "Foo"})
+             |> put_req_header("content-type", "application/graphql")
+             |> plug_parser
+             |> Absinthe.Plug.call(opts)
 
     assert resp_body == ~s({"data":{"user":"Foo"}})
   end
 
   test "context passed correctly to resolvers with documents provided as literals" do
-    opts = Absinthe.Plug.init(schema: TestSchema, document_providers: [__MODULE__.LiteralDocuments])
+    opts =
+      Absinthe.Plug.init(
+        schema: TestSchema,
+        document_providers: [Absinthe.Plug.TestLiteralDocuments]
+      )
 
-    assert %{status: 200, resp_body: resp_body} = conn(:post, "/", %{"id" => "2"})
-    |> Absinthe.Plug.put_options(context: %{user: "Foo"})
-    |> put_req_header("content-type", "application/graphql")
-    |> plug_parser
-    |> Absinthe.Plug.call(opts)
+    assert %{status: 200, resp_body: resp_body} =
+             conn(:post, "/", %{"id" => "2"})
+             |> Absinthe.Plug.put_options(context: %{user: "Foo"})
+             |> put_req_header("content-type", "application/graphql")
+             |> plug_parser
+             |> Absinthe.Plug.call(opts)
 
     assert resp_body == ~s({"data":{"user":"Foo"}})
   end
 
   test "works using documents loaded from an extracted_queries.json" do
-    opts = Absinthe.Plug.init(schema: TestSchema, document_providers: [__MODULE__.ExtractedDocuments])
+    opts =
+      Absinthe.Plug.init(schema: TestSchema, document_providers: [__MODULE__.ExtractedDocuments])
 
-    assert %{status: 200, resp_body: resp_body} = conn(:post, "/", %{"id" => "1", "variables" => %{"id" => "foo"}})
-    |> put_req_header("content-type", "application/graphql")
-    |> plug_parser
-    |> Absinthe.Plug.call(opts)
+    assert %{status: 200, resp_body: resp_body} =
+             conn(:post, "/", %{"id" => "1", "variables" => %{"id" => "foo"}})
+             |> put_req_header("content-type", "application/graphql")
+             |> plug_parser
+             |> Absinthe.Plug.call(opts)
 
     assert resp_body == @result
   end
 
   test ".get compiled" do
-    assert %Absinthe.Blueprint{} = Compiled.get(LiteralDocuments, "1")
-    assert %Absinthe.Blueprint{} = Compiled.get(LiteralDocuments, "1", :compiled)
+    assert %Absinthe.Blueprint{} = Compiled.get(Absinthe.Plug.TestLiteralDocuments, "1")
+
+    assert %Absinthe.Blueprint{} =
+             Compiled.get(Absinthe.Plug.TestLiteralDocuments, "1", :compiled)
   end
 
   test ".get source" do
-    assert @query == Compiled.get(LiteralDocuments, "1", :source)
+    assert @query == Compiled.get(Absinthe.Plug.TestLiteralDocuments, "1", :source)
   end
 
+  test "telemetry events executed", context do
+    :telemetry.attach_many(
+      context.test,
+      [
+        [:absinthe, :execute, :operation, :start],
+        [:absinthe, :execute, :operation, :stop]
+      ],
+      fn event, measurements, metadata, config ->
+        send(self(), {event, measurements, metadata, config})
+      end,
+      %{}
+    )
+
+    opts =
+      Absinthe.Plug.init(
+        schema: TestSchema,
+        document_providers: [Absinthe.Plug.TestLiteralDocuments]
+      )
+
+    assert %{status: 200, resp_body: resp_body} =
+             conn(:post, "/", %{"id" => "2"})
+             |> Absinthe.Plug.put_options(context: %{user: "Foo"})
+             |> put_req_header("content-type", "application/graphql")
+             |> plug_parser
+             |> Absinthe.Plug.call(opts)
+
+    assert_receive {[:absinthe, :execute, :operation, :start], _, %{id: id}, _config}
+    assert_receive {[:absinthe, :execute, :operation, :stop], measurements, %{id: ^id}, _config}
+
+    assert is_number(measurements[:duration])
+
+    :telemetry.detach(context.test)
+  end
 end